+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Benchmarks added
 - Export Key/Nonce/… constructors from Internal module
 - New password hashing module
+- Show instances for most (all?) relevant data types
 
 ## [0.1.1.1] - 2021-01-15
 ### Changed